+++ core/src/com/unciv/logic/city/CityConstructions.kt
@@ -1,7 +1,6 @@
 package com.unciv.logic.city
 
 import com.badlogic.gdx.graphics.Color
-import com.unciv.Constants
 import com.unciv.logic.automation.ConstructionAutomation
 import com.unciv.logic.civilization.AlertType
 import com.unciv.logic.civilization.PopupAlert
@@ -34,8 +33,6 @@ class CityConstructions {
 
     var builtBuildings = HashSet<String>()
     val inProgressConstructions = HashMap<String, Int>()
-    @Deprecated("As of 3.7.5, all constructions are in the queue")
-    var currentConstruction=""
     var currentConstructionFromQueue: String
         get() {
             if(constructionQueue.isEmpty()) return "" else return constructionQueue.first()
@@ -73,12 +70,10 @@ class CityConstructions {
             stats.add(building.getStats(cityInfo.civInfo))
 
         for (unique in builtBuildingUniqueMap.getAllUniques()) when (unique.placeholderText) {
-            "[] Per [] Population in this city" -> stats.add(unique.stats!!.times(cityInfo.population.population / unique.params[1].toFloat()))
-            "[] once [] is discovered" -> if (cityInfo.civInfo.tech.isResearched(unique.params[1])) stats.add(unique.stats!!)
+            "[] Per [] Population in this city" -> stats.add(unique.stats.times(cityInfo.population.population / unique.params[1].toFloat()))
+            "[] once [] is discovered" -> if (cityInfo.civInfo.tech.isResearched(unique.params[1])) stats.add(unique.stats)
         }
 
-        // This is to be deprecated and converted to "[stats] Per [N] Population in this city" - keeping it here to that mods with this can still work for now
-        stats.science += (builtBuildingUniqueMap.getAllUniques().count { it.text == "+1 Science Per 2 Population" } * cityInfo.population.population / 2).toFloat()
         return stats
     }
 
@@ -305,6 +300,9 @@ class CityConstructions {
                     civ.addNotification("[${construction.name}] has been built in [${cityInfo.name}]", cityInfo.location, Color.BROWN)
                 else
                     civ.addNotification("[${construction.name}] has been built in a faraway land",null,Color.BROWN)
+            }
+            if(construction.science > 0 && cityInfo.civInfo.hasUnique("Receive a tech boost when scientific buildings/wonders are built in capital")){
+                cityInfo.civInfo.tech.addScience(cityInfo.civInfo.tech.scienceOfLast8Turns.sum() / 8)
             }
         } else
             cityInfo.civInfo.addNotification("[${construction.name}] has been built in [" + cityInfo.name + "]", cityInfo.location, Color.BROWN)